made changes to init_rule in Parser class, added functions to Parser and Grammar, introduced ParseException, updated documentation for various classes, updated framework
Fixed init_rule, as it was bugged. Replaced
sentence_contains_unknown_words with get_unknown_words function, but
kept the other one as legacy method for now. Added a method
get_rules_via_lhs to the Grammar class. It is required by the
initialization rule, as the initial rules are picked via LHS, not RHS.
Also gave a suggestion for renaming get_possible_parent_rules, because I
always have to stop and think what it means exactly. Otherwise I wrote
some documentation in various classes, updated the framework with my
changes and added two things to the readme.